requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Django==5.2.13
greenlet==3.4.0
idna==3.11
numpy==2.2.5
oracledb==3.1.1
pandas==2.2.3
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.4.0
gunicorn==23.0.0